introduce utils.mjs
reduce code redundance
