+++ content/options.xhtml
@@ -4,7 +4,7 @@
     <meta charset="utf-8"/>
     <title data-l10n-content="options.title"></title>
     <link rel="stylesheet" href="options.css" type="text/css"/>
-    <script src="options.js" defer=""></script>
+    <script type="module" src="options.js"></script>
   </head>
   <body id="options" class="vbox">
     <div id="d_content" class="vbox">
